Add feature to confirm that user input is number
The javascript logic was modified to check that user input is a number upon clicking the convert button. If it is, the conversion will run and render in the appropriate divs. If not, text asking the user to please enter a number will be rendered where the converted figures would otherwise be.